+++ src/_packages/core.index.ts
@@ -1,4 +1,3 @@
-import { PromptbookLibrary } from '../classes/PromptbookLibrary';
 import { promptbookStringToJson } from '../conversion/promptbookStringToJson';
 import { validatePromptbookJson } from '../conversion/validatePromptbookJson';
 import { createPromptbookExecutor } from '../execution/createPromptbookExecutor';
@@ -6,11 +5,12 @@ import { MockedEchoNaturalExecutionTools } from '../execution/plugins/natural-ex
 import { CallbackInterfaceTools } from '../execution/plugins/user-interface-execution-tools/callback/CallbackInterfaceTools';
 import { CallbackInterfaceToolsOptions } from '../execution/plugins/user-interface-execution-tools/callback/CallbackInterfaceToolsOptions';
 import { SimplePromptInterfaceTools } from '../execution/plugins/user-interface-execution-tools/simple-prompt/SimplePromptInterfaceTools';
+import { SimplePromptbookLibrary } from '../library/SimplePromptbookLibrary';
 import { ExecutionTypes } from '../types/ExecutionTypes';
 import { PROMPTBOOK_VERSION } from '../version';
 
 // @promptbook/core
-export { ExecutionTypes, PROMPTBOOK_VERSION, PromptbookLibrary };
+export { ExecutionTypes, PROMPTBOOK_VERSION, SimplePromptbookLibrary };
 
 // @promptbook/simple-prompt
 export { SimplePromptInterfaceTools };